+++ internal/pkg/clients/types_test.go
@@ -19,136 +19,163 @@ import (
 	"regexp"
 	"testing"
 
+	clientsimages "github.com/NethermindEth/sedge/configs/images"
 	"github.com/stretchr/testify/assert"
 )
 
+var guesser clientsimages.ClientsImagesGuesser
+
 func TestSetImageOrDefault_Execution(t *testing.T) {
+	// default clients images
+	defaults, err := clientsimages.NewDefaultClientsImages()
+	if err != nil {
+		t.Fatal(err)
+	}
+
 	tests := []struct {
 		client        Client
-		expectedImage regexp.Regexp
+		expectedImage *regexp.Regexp
 	}{
 		{
 			client: Client{
 				Name: "geth",
 				Type: "execution",
 			},
-			expectedImage: *regexp.MustCompile(`^ethereum/client-go:v\d+\.\d+\.\d+$`),
+			expectedImage: guesser.GethRegexp(),
 		},
 		{
 			client: Client{
 				Name: "besu",
 				Type: "execution",
 			},
-			expectedImage: *regexp.MustCompile(`^hyperledger/besu:\d+\.\d+\.\d+$`),
+			expectedImage: guesser.BesuRegexp(),
 		},
 		{
 			client: Client{
 				Name: "nethermind",
 				Type: "execution",
 			},
-			expectedImage: *regexp.MustCompile(`^nethermind/nethermind:\d+\.\d+\.\d+$`),
+			expectedImage: guesser.NethermindRegexp(),
 		},
 		{
 			client: Client{
 				Name: "erigon",
 				Type: "execution",
 			},
-			expectedImage: *regexp.MustCompile(`^thorax/erigon:v\d+\.\d+\.\d+$`),
+			expectedImage: guesser.ErigonRegexp(),
 		},
 	}
 	for _, test := range tests {
 		t.Run(test.client.Name, func(t *testing.T) {
-			test.client.SetImageOrDefault("")
-			assert.True(t, test.expectedImage.Match([]byte(test.client.Image)))
+			test.client.SetImageOrDefault("", defaults)
+			assert.Regexp(t, test.expectedImage, test.client.Image)
 		})
 	}
 }
 
 func TestSetImageOrDefault_Consensus(t *testing.T) {
+	// default clients images
+	defaults, err := clientsimages.NewDefaultClientsImages()
+	if err != nil {
+		t.Fatal(err)
+	}
+
 	tests := []struct {
 		client        Client
-		expectedImage regexp.Regexp
+		expectedImage *regexp.Regexp
 	}{
 		{
 			client: Client{
 				Name: "lighthouse",
 				Type: "consensus",
 			},
-			expectedImage: *regexp.MustCompile(`^sigp/lighthouse:v\d+\.\d+\.\d+$`),
+			expectedImage: guesser.LighthouseConsensusRegexp(),
 		},
 		{
 			client: Client{
 				Name: "prysm",
 				Type: "consensus",
 			},
-			expectedImage: *regexp.MustCompile(`^gcr.io/prysmaticlabs/prysm/beacon-chain:v\d+\.\d+\.\d+$`),
+			expectedImage: guesser.PrysmConsensusRegexp(),
 		},
 		{
 			client: Client{
 				Name: "teku",
 				Type: "consensus",
 			},
-			expectedImage: *regexp.MustCompile(`^consensys/teku:\d+\.\d+\.\d+$`),
+			expectedImage: guesser.TekuConsensusRegexp(),
 		},
 		{
 			client: Client{
 				Name: "lodestar",
 				Type: "consensus",
 			},
-			expectedImage: *regexp.MustCompile(`^chainsafe/lodestar:v\d+\.\d+\.\d+$`),
+			expectedImage: guesser.LodestarConsensusRegexp(),
 		},
 	}
 	for _, test := range tests {
 		t.Run(test.client.Name, func(t *testing.T) {
-			test.client.SetImageOrDefault("")
-			assert.True(t, test.expectedImage.Match([]byte(test.client.Image)))
+			test.client.SetImageOrDefault("", defaults)
+			assert.Regexp(t, test.expectedImage, test.client.Image)
 		})
 	}
 }
 
 func TestSetImageOrDefault_Validator(t *testing.T) {
+	// default clients images
+	defaults, err := clientsimages.NewDefaultClientsImages()
+	if err != nil {
+		t.Fatal(err)
+	}
+
 	tests := []struct {
 		client        Client
-		expectedImage regexp.Regexp
+		expectedImage *regexp.Regexp
 	}{
 		{
 			client: Client{
 				Name: "lighthouse",
 				Type: "validator",
 			},
-			expectedImage: *regexp.MustCompile(`^sigp/lighthouse:v\d+\.\d+\.\d+$`),
+			expectedImage: guesser.LighthouseValidatorRegexp(),
 		},
 		{
 			client: Client{
 				Name: "prysm",
 				Type: "validator",
 			},
-			expectedImage: *regexp.MustCompile(`^gcr.io/prysmaticlabs/prysm/validator:v\d+\.\d+\.\d+$`),
+			expectedImage: guesser.PrysmValidatorRegexp(),
 		},
 		{
 			client: Client{
 				Name: "teku",
 				Type: "validator",
 			},
-			expectedImage: *regexp.MustCompile(`^consensys/teku:\d+\.\d+\.\d+$`),
+			expectedImage: guesser.TekuValidatorRegexp(),
 		},
 		{
 			client: Client{
 				Name: "lodestar",
 				Type: "validator",
 			},
-			expectedImage: *regexp.MustCompile(`^chainsafe/lodestar:v\d+\.\d+\.\d+$`),
+			expectedImage: guesser.LodestarValidatorRegexp(),
 		},
 	}
 	for _, test := range tests {
 		t.Run(test.client.Name, func(t *testing.T) {
-			test.client.SetImageOrDefault("")
+			test.client.SetImageOrDefault("", defaults)
 			assert.True(t, test.expectedImage.Match([]byte(test.client.Image)))
 		})
 	}
 }
 
 func TestSetImageOrDefault_CustomImage(t *testing.T) {
+	// default clients images
+	defaults, err := clientsimages.NewDefaultClientsImages()
+	if err != nil {
+		t.Fatal(err)
+	}
+
 	tests := []struct {
 		client      Client
 		customImage string
@@ -177,7 +204,7 @@ func TestSetImageOrDefault_CustomImage(t *testing.T) {
 	}
 	for _, test := range tests {
 		t.Run(test.client.Name, func(t *testing.T) {
-			test.client.SetImageOrDefault(test.customImage)
+			test.client.SetImageOrDefault(test.customImage, defaults)
 			assert.Equal(t, test.customImage, test.client.Image)
 		})
 	}